Register output handler to sync with commander
Currently the terminal component is unaware that its content has been
updated if the history addon is used. So instead of waiting for the
onData callback to trigger, we register an output listener that fires
anytime the output changes. We can then properly keep the output synced
between terminal and commander.