Fix: Allow files paths with spaces
Without quotes around `${filepath}` the command fails e.g. if the file
path contains spaces, which is already true for the default path where
the app stores PCAP on macOS.

The quotes around `${pathToWireshark}` are also changed to single quotes
so that some special characters like `$` are not interpreted in the case
that they happen to be in `pathToWireshark`.