Fix running the correct executable on macOS
Finding the correct executable on macOS is a bit tricky:
The application is e.g. /Applications/Wireshark.app, while
the executable is e.g.
/Applications/Wireshark.app/Contents/MacOS/Wireshark.

When people select a Wireshark on macOS, they will select something like
`/Applications/Wireshark.app`, because in the Finder people only see the
apps, it is hard to select files _within_ an app and because we also ask
in`askForWiresharkExecutable` to select something with the extension
`app`.

So what we have to do on macOS: Store the path to the app (e.g.
`/Applications/Wireshark.app`) and then when we want to run the
executable concat the path to the executable (usually
`Contents/MacOS/Wireshark`) within the app to the app path.