Trying to make the automated build more reliable
We are currently using our own internal npm registry on this project.
That registry is occasionally unreliable, so `npm ci` sometimes fails
because it fails to find some needed packages on the server.

When `npm ci` fails, this commits retries it up to three times. This is
a bit desperate but seems like the best compromise for now.