Unify naming
`pathToWireshark`, `wiresharkLocation`, and `wiresharkPath` were all
used, with some variations on these, without there being a different
meaning between the different names.

So this is unified to `wiresharkPath`. Sometimes with a prefix (e.g.
`selectedWiresharkPath`) to make it more meaningful and sometimes
dropping the `wireshark` if the context already makes it clear that it
is about Wireshark.