Change semantic of dbFilePath
For the upcoming automatic selection of trace DBs I want to model it
explicitly whether users have not set a trace DBs manually (which now
means that dbFilePath is undefined) or if they did (then it still
contains that path).

This leads to a small change in behaviour: If users now select the
default DB manually, it previously seemed like they did not select a
trace DB manually. This has now changed. But this will not matter in the
future anyway, because there will not be a default trace DB anymore.