Switch to selectedDevice from shared
Directly reading the state properties is either fragile, because it
would break whenever we change the shape of the state or inflexible if
it would hinder us to do any changes to the shape of the state.

Thus instead of reading the state directly it is better to use a
selector provided by the module that is also responsible for the shape
of the state slice. Luckily such a selector already exists:
selectedDevice. It is just not exported by shared yet.