Add +CEREG read processor and update associated tests
The insight into how +CEREG responses work in practice is not, with
regards to what kind of responses are likely to appear to the customer,
is not very clear to the writer at the moment.

Therefore, I expect we may need to adjust the processor once we get
more manual testing by internal, and potentially external, customers.