UX improvement by reducing number of clicks
In order for the user to Open Terminal, Start Tracing, and sending the
recommended AT commands, a lot of clicks need to be performed. This
commit aims at a small improvement on the matter.

When the user first connects to a device, we will automatically select
the first VComIndex port as the one to be used over UART in order to
communicate with e.g. the AT_Client.

Also adds an option to manually select the serial port to be used when
opening serial terminal and sending recommended AT commands.