Give load button a proper loading state
When reading a file, since it may take a bit of time now that tshark is
always run in the background, it would be helpful to properly display
that we are reading the file.

This commit disables the read file button, while it's reading, and also
gives the button the standard snake active animation.