change: progress CB to always throttle on duration
To convert a reasonably small binary file of 3.8MB, you'd receive well
over 100K progress callbacks. Updating the UI on each CB is not
feasible, and will quickly start to slow the app and block the user.

Before this commit, throttling was possible, but it was an arbitrary
interval of progress CBs that was handled, and even with this the UI was
yet observed to block.

Even more, if we'd increase the interval, the Trace Details > File size
would report wrong. This is because it could miss X number of CBs in the
end. Hence the motivation to use an interval to process the latest progress.