feat(AT): Send given related AT command
In each DashboardCard field, if the related command has a recommended at
command, display the `Run` option in the tooltip. If the user clicks the
`Run` option, run the recommended command if uartSerialPort is open.

This enables users to quickly get an overview of how they may populate
the fields in the dashboard.