fix(DashboardState): missing last packet in dashboard update
When a new timestamp is set the useEffect to update the dashboard state
is run. Inside it, we have a filter in case user has paned backward in
time and we don't want to display the latest state.

The filter was not inclusive on the timstamp, and hence always filtered
the last packet. This commit makes it an inclusive filter.

In addition, since the timestamp is updated for very frequently in some
cases, we don't want to run though `events.reduce()` for each updated
timstamp. Therefore, we've added a `setTimeout`, and a `clearTimeout`.

The setTimeout will put each of the tasks in the microQueue, and because
of how it works, if several `events.reduce()` would be queued, only the
last one will run.