fix: TX power reduction processor only partially worked
AT parsing removed all \r\n, which made it difficult or impossible to
separate the return values from the modem. In the case of %XEMPR read,
one line is given for LTE-M and one line for NB-IoT.

This commit makes sure that AT processors are able to split lines in
such cases, and control whether or not to separate them.