Add more efficient auto detect db method
The current auto detection may take some time to return a version, or
may fail. By using the AT command AT+CGMR, we may receive the modem
version almost instantly.

Drawback is that we need to parse the AT command directly from the
uartPort, instead of getting it from trace. This is because we need the
database before we can start receiving data from trace.