Forcefully empty shell buffer, in order to detect db version
This is one of several approaches, but in order to make it 'just work',
this may be a good option. The call to `shellParser.unPause()`, will
remove anything that the user has entered into the shell.

An alternative could be to skip auto detecting db version via `at+cgmr`,
if the shell is paused.