Add toggle inside the dialog
It's easier for the user to to understand how to turn of `Open in
Wireshark` if the option to disable it is in the Dialog itself.

Additionally, when toggling it in the Dialog, the toggle in the Side
Menu is also visibly being toggled, so that it may educate the user.