change: Connection Status
Earlier version had timeout on all of the statuses, which would give
them a failed state in the GUI. This commit omits most of the timeouts,
and only keeps a timeout on Trace.

For an improved UI we omit success caption, and only display the green
check mark. Also, we try to derive the failed captions from the actual
cause or status message provided by e.g. AT commands.