fix: setDashboardState persisted state from future
When paning the event chart backwards in time, we should not persist the
events that has happened in the "future". If a field was set to
undefined, but was defined before, it would not be overwritten.

Since we always run through all the events before setting the new
dashboard state, there's no reason to also deconstruct the old state.
This commit only sets the new state based on the reduced new state.