change: the order of connection status
To easier check the previous step, which may affect the next, it may
be more suitable to verify the last step first, and then work backwards.

The assumption in this commit, is that if modem has established a
connection to an APN, then it must have successfully established a
connection to LTE, and SIM card must be initialized.

Similarily, if LTE Connection is established, the Sim card must have
been successfully initialized.