feat: Improve operator name formatting
Observed that operator may be an empty string. When this happens the
field is empty. In this case, the field should not be empty, and it
should not be 'Unknown'.

This commit checks if both operator variables are empty strings, and if
they are the operator field shows 'Not Available'.