Reset state when device is selected
Resets the dashboard state, relevant trace information, and clears the
events from the chart.

Works well after e.g. a file is completely read, but if device is
selected while reading a file, it clears everything, but continues to
read the file. Not a priority to fix now, but should be fixed later.