Store what value to read network status from
Use both conevalResult and networkStatus in order to decide the status
of the LTE connection

In order to read the most valid network status, we set what value is
last updated in the dashboard state. In the connection status component
we may then more accurately decide what value to validate.