Terminal Serial Port improvements
Only auto-connect to the terminal serial port if the device have exposed
two or more serial ports. If only one serial port is exposed, then we
expect that to be used for modem trace.

Extract redux state to separate Slice containing the "terminal
serial port" and the shell parser.

Add more verbose debug logging in order to improve debugging of terminal serial
port connection. We may use these to ask users what is expected, and compare to
what the app is able to detect.