Improve signal quality values
Fix: when the index for rsrp, rsrq, and snr, are 255, then the related
decibel value should not be calculated, and in the GUI it should display
'Not known or not detectable', as per the documentation.

Feature: When decibel value is available, display the threshold values
Excellent, Good, Fair, Poor, and the decibel value in parenthesis.
Example (rsrp), 'Excellent (-80 dBm)'.