fix: Do not update chart if there are not packets
If there were no packets in list of TraceEvents, then the panZoomPlugin
would try to access timestamp on an undefined value, and would throw an
uncaught expection.

This would result in the closeDevice function not to properly run, and
hence restrict device to be deselected. This commit ensures that the
stop-process event will return early if there are no EventPackets.