fix: make log messages more intuitive
When opening Cellular Monitor, and reading that logging is enabled
with log level "0", makes it sound like the user should expect to get
logging from nrf-monitor-lib.

Make it more understandable when the logging is off, and when it is on,
so that when the user actually turns it on, then the user can be
confident that the logging is on.