When loading trace databases default to AllDevices
Which means that if a specific chip is detected, it will
return all trace databases. It makes more sense to present
all choices, rather than no choice.

Works well when user wants to open a file, and needs to
select the appropriate trace database.

Fixes the issue where it would log errors because device
was undefined in `getNrfDeviceVersion`.