Add v2.0.0 trace database
Change device key from nRF9161 to nRF91x1, in order to make it
generic for nRF9131 and nRF9161.

The file name for v2.0.0 has been altered because of a rushed
decision when we released v2.0.0-beta.

We cannot simply swap v2.0.0 -> v2.0.0-beta, because then our
program would think that old v2.0.0-beta files would be the
official v2.0.0, and skip downloading the new files.